rename FahrzeugPanel to FahrzeugPanelMessstelle, format Dates in Tooltip
