+++ frontend/components.d.ts
@@ -30,8 +30,9 @@ declare module '@vue/runtime-core' {
     DeleteDialog: typeof import('./src/components/pdfreport/assetforms/DeleteDialog.vue')['default']
     DetektierteFahrzeugartIcon: typeof import('./src/components/messstelle/DetektierteFahrzeugartIcon.vue')['default']
     DisplayImageAsset: typeof import('./src/components/pdfreport/assets/DisplayImageAsset.vue')['default']
-    FahrzeugPanel: typeof import('./src/components/messstelle/optionsmenue/panels/FahrzeugPanel.vue')['default']
+    FahrzeugPanel: typeof import('./src/components/zaehlstelle/optionsmenue/panels/FahrzeugPanel.vue')['default']
     FahrzeugPanelFahrzeugkategorienContent: typeof import('./src/components/messstelle/optionsmenue/panels/content/FahrzeugPanelFahrzeugkategorienContent.vue')['default']
+    FahrzeugPanelMessstelle: typeof import('./src/components/messstelle/optionsmenue/panels/FahrzeugPanelMessstelle.vue')['default']
     FahrzeugPanelVerkehrsartenContent: typeof import('./src/components/messstelle/optionsmenue/panels/content/FahrzeugPanelVerkehrsartenContent.vue')['default']
     FilterOptionen: typeof import('./src/components/messstelle/optionsmenue/FilterOptionen.vue')['default']
     GeometriePanel: typeof import('./src/components/zaehlstelle/optionsmenue/panels/GeometriePanel.vue')['default']
@@ -87,9 +88,12 @@ declare module '@vue/runtime-core' {
     VExpansionPanelHeader: typeof import('vuetify/lib')['VExpansionPanelHeader']
     VExpansionPanels: typeof import('vuetify/lib')['VExpansionPanels']
     VFadeTransition: typeof import('vuetify/lib')['VFadeTransition']
+    VFileInput: typeof import('vuetify/lib')['VFileInput']
     VFooter: typeof import('vuetify/lib')['VFooter']
+    VForm: typeof import('vuetify/lib')['VForm']
     VHover: typeof import('vuetify/lib')['VHover']
     VIcon: typeof import('vuetify/lib')['VIcon']
+    VImg: typeof import('vuetify/lib')['VImg']
     VList: typeof import('vuetify/lib')['VList']
     VListItem: typeof import('vuetify/lib')['VListItem']
     VListItemAction: typeof import('vuetify/lib')['VListItemAction']
@@ -114,7 +118,9 @@ declare module '@vue/runtime-core' {
     VTabItem: typeof import('vuetify/lib')['VTabItem']
     VTabs: typeof import('vuetify/lib')['VTabs']
     VTabsItems: typeof import('vuetify/lib')['VTabsItems']
+    VTextarea: typeof import('vuetify/lib')['VTextarea']
     VTextField: typeof import('vuetify/lib')['VTextField']
+    VToolbar: typeof import('vuetify/lib')['VToolbar']
     VToolbarTitle: typeof import('vuetify/lib')['VToolbarTitle']
     VTooltip: typeof import('vuetify/lib')['VTooltip']
     WetterIcon: typeof import('./src/components/zaehlstelle/icons/WetterIcon.vue')['default']
