+++ frontend/src/components/map/ZaehlstelleMap.vue
@@ -146,6 +146,7 @@ import markerIconDiamondShadow from "@/assets/cards-diamond-shadow.png";
 import TooltipMessstelleDTO from "@/types/TooltipMessstelleDTO";
 import AnzeigeKarteDTO from "@/types/AnzeigeKarteDTO";
 import MessstelleKarteDTO from "@/types/MessstelleKarteDTO";
+import { useDateUtils } from "@/util/DateUtils";
 
 @Component({
     components: {
@@ -509,13 +510,19 @@ export default class ZaehlstelleMap extends Vue {
             tooltip = `${tooltip}${tooltipDto.stadtbezirk}<br/>`;
         }
         if (tooltipDto.realisierungsdatum) {
-            tooltip = `${tooltip} Aufbau: ${tooltipDto.realisierungsdatum}<br/>`;
+            tooltip = `${tooltip} Aufbau: ${useDateUtils().formatDate(
+                tooltipDto.realisierungsdatum
+            )}<br/>`;
         }
         if (tooltipDto.abbaudatum) {
-            tooltip = `${tooltip}Abbau: ${tooltipDto.abbaudatum}<br/>`;
+            tooltip = `${tooltip}Abbau: ${useDateUtils().formatDate(
+                tooltipDto.abbaudatum
+            )}<br/>`;
         }
         if (tooltipDto.datumLetztePlausibleMessung) {
-            tooltip = `${tooltip}Letzte plausible Messung: ${tooltipDto.datumLetztePlausibleMessung}<br/>`;
+            tooltip = `${tooltip}Letzte plausible Messung: ${useDateUtils().formatDate(
+                tooltipDto.datumLetztePlausibleMessung
+            )}<br/>`;
         }
 
         tooltip = `${tooltip}</div>`;
