+++ frontend/src/components/messstelle/optionsmenue/panels/FahrzeugPanelMessstelle.vue
